[Refactor] #92 - 토스트 메서드 위치 DSKit으로 변경
